requirements.txt
accelerate==1.6.0
black==25.1.0
byted_omnistore==1.0.2
cloudpickle==3.1.1
decord==0.6.0
diffusers==0.33.1
einops==0.8.1
flash_attn==2.6.3
flashattn_hopper==3.0.0b1
goku==0.2.5
hydra-core==1.3.2
imageio==2.35.1
numpy==2.2.5
omegaconf==2.3.0
openai==1.78.1
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
pandas==2.2.3
Pillow==11.2.1
proton==0.9.1
PyYAML==6.0.2
Requests==2.32.3
safetensors==0.5.3
torch==2.5.1
torchvision==0.20.1
transformers==4.51.3